feature: handling w3c and oa documents
