+++ src/templates/govtech-demo-cert/transcript.tsx
@@ -1,256 +1,23 @@
-import React, { FunctionComponent, useState } from "react";
-import { get } from "lodash";
-import { formatDate } from "./common/functions";
-import "bootstrap/dist/css/bootstrap.css";
-import "./common/demo-styles.css";
-import { ObfuscatableValue, TemplateProps } from "@govtechsg/decentralized-renderer-react-components";
+import React, { FunctionComponent } from "react";
+import { TemplateProps } from "@govtechsg/decentralized-renderer-react-components";
 import { GovtechOpencertsTemplateCertificate } from "../samples";
-import { PrintWatermark } from "./common/print-watermark";
-import govtechLogo from "./common/govtech-logo.png";
-import transcriptBg from "./common/transcript-background.png";
-import { SimplePrivacyFilterBanner } from "./common/simple-privacy-filter-banner";
-
-/** True if document is a W3C Verifiable Credential (not OpenAttestation v3/v4). */
-function isW3CDocument(document: Record<string, unknown>): boolean {
-  if (!document || !document["@context"]) return false;
-  const type = document.type;
-  const hasVerifiableCredential = Array.isArray(type)
-    ? type.includes("VerifiableCredential")
-    : type === "VerifiableCredential";
-  if (!hasVerifiableCredential) return false;
-  if (document.openAttestationMetadata) return false;
-  const ctx = document["@context"];
-  if (
-    (Array.isArray(ctx) && ctx.some(c => c && String(c).includes("openattestation"))) ||
-    (typeof ctx === "string" && ctx.includes("openattestation"))
-  ) {
-    return false;
-  }
-  return true;
-}
+import { TranscriptTemplate as W3cTranscriptTemplate } from "./w3c/w3cTranscript";
+import { GovtechW3cVerifiableCredential } from "../samples/govtech-w3c-teamplate-sample";
+import { OpenAttestationTranscript } from "./openattestation/OpenAttestationTranscript";
+import { isW3CDocument } from "../../utils/w3c-utils";
 
 export const TranscriptTemplate: FunctionComponent<TemplateProps<GovtechOpencertsTemplateCertificate>> = ({
   document,
   handleObfuscation
 }) => {
-  const [editable, setEditable] = useState(false);
-  const isW3C = isW3CDocument((document as unknown) as Record<string, unknown>);
-  const documentName = get(document, "name");
-  const documentId = get(document, "id");
-  const issuanceDate = get(document, "issuedOn");
-  const admissionDate = get(document, "admissionDate");
-  const graduationDate = get(document, "graduationDate");
+  if (isW3CDocument(document as unknown as Record<string, unknown>)) {
+    return (
+      <W3cTranscriptTemplate
+        document={document as unknown as GovtechW3cVerifiableCredential}
+        handleObfuscation={handleObfuscation}
+      />
+    );
+  }
 
-  const recipientName = get(document, "recipient.name");
-  const recipientNric = get(document, "recipient.nric");
-  const recipientCourse = get(document, "recipient.course");
-  const studentId = get(document, "additionalData.studentId");
-
-  const transcriptData = document.transcript || [];
-
-  const transcriptSection = transcriptData.map((t, i) => (
-    <tr key={i}>
-      <td>
-        {isW3C ? (
-          t.courseCode
-        ) : (
-          <ObfuscatableValue
-            editable={editable}
-            value={t.courseCode}
-            onObfuscationRequested={() => handleObfuscation(`transcript[${i}].courseCode`)}
-          />
-        )}
-      </td>
-      <td>
-        {isW3C ? (
-          t.name
-        ) : (
-          <ObfuscatableValue
-            editable={editable}
-            value={t.name}
-            onObfuscationRequested={() => handleObfuscation(`transcript[${i}].name`)}
-          />
-        )}
-      </td>
-      <td>
-        {isW3C ? (
-          t.grade
-        ) : (
-          <ObfuscatableValue
-            editable={editable}
-            value={t.grade}
-            onObfuscationRequested={() => handleObfuscation(`transcript[${i}].grade`)}
-          />
-        )}
-      </td>
-      <td>
-        {isW3C ? (
-          t.courseCredit
-        ) : (
-          <ObfuscatableValue
-            editable={editable}
-            value={t.courseCredit}
-            onObfuscationRequested={() => handleObfuscation(`transcript[${i}].courseCredit`)}
-          />
-        )}
-      </td>
-      <td>
-        {isW3C ? (
-          t.semester
-        ) : (
-          <ObfuscatableValue
-            editable={editable}
-            value={t.semester}
-            onObfuscationRequested={() => handleObfuscation(`transcript[${i}].semester`)}
-          />
-        )}
-      </td>
-    </tr>
-  ));
-
-  return (
-    <>
-      <PrintWatermark />
-      <div className="container">
-        {!isW3C && (
-          <SimplePrivacyFilterBanner onToggleEditable={() => setEditable(!editable)} className="privacy-banner" />
-        )}
-        <div
-          className="p-2 container"
-          style={{
-            backgroundImage: `url('${transcriptBg}')`,
-            backgroundRepeat: "repeat"
-          }}
-        >
-          <div className="row root cert-title" style={{ paddingLeft: "3%" }}>
-            <b>{documentName}</b>
-          </div>
-
-          <div
-            className="row transcript"
-            style={{
-              paddingTop: "3%",
-              paddingLeft: "2%"
-            }}
-          >
-            <div className="col">
-              <div className="row">
-                <div className="col">NAME</div>
-                <div className="col">
-                  :&nbsp;&nbsp;
-                  {recipientName}
-                </div>
-              </div>
-              <div className="row">
-                <div className="col">COURSE</div>
-                <div className="col">
-                  :&nbsp;&nbsp;
-                  {recipientCourse}
-                </div>
-              </div>
-              <div className="row">
-                <div className="col">NRIC/FIN</div>
-                <div className="col">
-                  :&nbsp;&nbsp;
-                  {recipientNric}
-                </div>
-              </div>
-              <div className="row">
-                <div className="col">STUDENT ID</div>
-                <div className="col">
-                  :&nbsp;&nbsp;
-                  {studentId}
-                </div>
-              </div>
-            </div>
-            <div className="col">
-              <div className="row">
-                <div className="col">DOCUMENT ID</div>
-                <div className="col">
-                  :&nbsp;&nbsp;
-                  {documentId}
-                </div>
-              </div>
-              <div className="row">
-                <div className="col">DATE OF ISSUANCE</div>
-                <div className="col">
-                  :&nbsp;&nbsp;
-                  {formatDate(issuanceDate)}
-                </div>
-              </div>
-              <div className="row">
-                <div className="col">DATE OF ADMISSION</div>
-                <div className="col">
-                  :&nbsp;&nbsp;
-                  {formatDate(admissionDate)}
-                </div>
-              </div>
-              <div className="row">
-                <div className="col">DATE OF GRADUATION</div>
-                <div className="col">
-                  :&nbsp;&nbsp;
-                  {formatDate(graduationDate)}
-                </div>
-              </div>
-            </div>
-          </div>
-
-          {transcriptData.length > 0 && (
-            <div className="row mb-4" style={{ paddingLeft: "3%", paddingTop: "5%" }}>
-              <div className="root cert-title">
-                <b>Transcript</b>
-              </div>
-              <table className="w-100 transcript">
-                <tbody>
-                  <tr>
-                    <th>Course Code</th>
-                    <th>Name</th>
-                    <th>Grade</th>
-                    <th>Units</th>
-                    <th>Semester</th>
-                  </tr>
-                  {transcriptSection}
-                </tbody>
-              </table>
-            </div>
-          )}
-
-          <div className="row">
-            <div className="col">
-              <img
-                className="w-100"
-                style={{
-                  paddingTop: "40%",
-                  paddingLeft: "3%",
-                  width: "100%",
-                  height: "auto"
-                }}
-                src={govtechLogo}
-                alt="Govtech Logo"
-              />
-            </div>
-            <div className="col" />
-            <div
-              className="col text-center"
-              style={{
-                paddingTop: "5%",
-                paddingRight: "5%",
-                width: "100%",
-                height: "auto"
-              }}
-            >
-              <img className="w-100" src={get(document, "additionalData.certSignatories[0].signature")} />
-              <hr className="m-1" />
-              <div className="transcript">
-                <b>{get(document, "additionalData.certSignatories[0].name")}</b>
-                <br />
-                {get(document, "additionalData.certSignatories[0].position")},{" "}
-                {get(document, "additionalData.certSignatories[0].organisation")}
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </>
-  );
+  return <OpenAttestationTranscript document={document} handleObfuscation={handleObfuscation} />;
 };
