+++ src/templates/govtech-demo-cert/w3c/w3cMedia.tsx
@@ -1,0 +1,18 @@
+import React, { FunctionComponent } from "react";
+import { TemplateProps } from "@govtechsg/decentralized-renderer-react-components";
+import "bootstrap/dist/css/bootstrap.css";
+import { GovtechW3cVerifiableCredential } from "../../samples/govtech-w3c-teamplate-sample";
+
+export const MediaTemplate: FunctionComponent<TemplateProps<GovtechW3cVerifiableCredential>> = () => {
+  return (
+    <div className="embed-responsive embed-responsive-16by9">
+      <iframe
+        id="youtube-vid"
+        className="embed-responsive-item"
+        src="https://www.youtube.com/embed/oskddwGpwUw?autoplay=1"
+        allowFullScreen
+      />
+    </div>
+  );
+};
+
