+++ src/templates/samples/index.tsx
@@ -1,3 +1,4 @@
 export * from "./govtech-template-sample";
 export * from "./testimonial-sample";
 export * from "./certificate-of-award-sample";
+export * from "./govtech-w3c-teamplate-sample";
